% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cutoff.R
\name{set_cutoff}
\alias{set_cutoff}
\title{Generate dataframe for use in vimes_multi when user specifying thresholds}
\usage{
set_cutoff(d_type, g1g1_threshold, mixed_threshold, g2g2_threshold)
}
\arguments{
\item{d_type}{one of 'temporal' or 'spatial'}

\item{g1g1_threshold}{cutoff distance for group 1 to group 1 transmission}

\item{mixed_threshold}{cutoff distance for mixed transmission}

\item{g2g2_threshold}{cutoff distance for group 2 to group 2 transmission}
}
\value{
a dataframe of 3 columns and 3 rows suitable for use in vimes_multi
}
\description{
Generate dataframe for use in vimes_multi when user specifying thresholds
}
\examples{
set_cutoff("spatial", 4,4,4)
}
