% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimes_prune_multi.R
\name{vimes_prune_multi}
\alias{vimes_prune_multi}
\title{Prune the graphs using group-specific values}
\usage{
vimes_prune_multi(
  x,
  cutoff = NULL,
  group_vect,
  graph_opt = vimes_graph_opt(),
  ...
)
}
\arguments{
\item{x}{A vimes data object}

\item{cutoff}{The cut-off values to be used for pruning}

\item{group_vect}{Vector of groups with same length as observations}

\item{graph_opt}{The settings for the graph. Imported from vimes.}

\item{...}{Any additional parameters relating to graph options}
}
\value{
list containing the pruned graph, cluster definition, and cutoff values used
}
\description{
Prune the graphs using group-specific values
}
\examples{
g1_obs <- 300 # no of cases observed for species 1
g2_obs <- 200 # no of cases observed for species 2
obs <- g1_obs + g2_obs
group_vect <- as.factor(sample(c(rep("g1", g1_obs), rep("g2", g2_obs)), replace = FALSE))

dat_time <- round(runif(obs, min = 0, max = 400)) #days from start of study
time_dist <- dist(dat_time)

dat_easting <- runif(obs, min = 372900, max = 656900) # this represents the distance data.
dat_northing <- runif(obs, min = 8744900, max = 9075900) # this represents the distance data.

dat_dist <- as.matrix(c(dat_easting, dat_northing))
dat_dist <- fields::rdist(dat_dist)/1000 # convert to km

x <- list(time_dist, dat_dist)
x <- vimes::vimes_data(x)
x <- x[[1]]
cutoff <- as.list(c(142, 143, 143))

vimes_prune_multi(
x=x,
cutoff = cutoff,
group_vect = group_vect,
graph_opt = vimes::vimes_graph_opt()
)

}
